requirements.txt
click
ffmpeg-python
playwright
requests
Telethon
python-dotenv
PyYAML
faster-whisper
numpy

